Add massive Controllers.h changes
First, the left and right chassis motors controlled by Controllers.c can now be set in the Motor and Sensors menu; This makes it possible to add multiple motors on one side.
Second, the vertical axis for the controllers has been scaled 1.3x to make driving straighter significantly easier with only a minute change in turning. The byte does not overflow as it is capped with motor_clamp