Fix Controller Module stop() bug
For some reason, stopping the current controller did not work, at all. The reason for this is because taskGetState did not return an expected value to check if the task was even running. To fix this, simply set the task handle to NULL if the current controller is not running and check that to see if the controller is running.