Upgrade to Pros 3.x
We finally got the V5 system; Yay!!! To celebrate, I've completely
upgraded and modified InuLib to work with the V5 system through the
inclusion of "api.h", "pros/api_legacy.h", "main.h", and the pros::c
namespace. All original PROS 2.x code

It is important to note that Serial.h and Serial.c and its use in other
modules had to be completely discarded as the V5 system does not have
UART or file-read-write ports at all; A small price to pay for
greatness.