Fix potentially dangereous bug in PIDMotor
When a instance of PIDMotor is created, it's target position is set to 0
and the system tries to make the motor go there immediantly. Thing is,
the motor might have already turned significantly and so this might
force the motor to power on for a long time, maybe damaging components
of the robot.

The current workaround is to only update the motor an instance of
PIDMotor is controlling iff the target position of PIDMotor was modified
at least once, thus forcing the user to make a delibrate decision