Separate Chassis Constructor code to Builders
Due to the handful of options a chassis can be configured with, making a
Builder for each chassis seemed like a good choice. Most of the code is
in inu/auto/chassis/AutoChassisBuilder.hpp, with the actual motors and
building segment in inu/auto/chassis/AutoXChassisBuilder.hpp.

The one drawback of this is that is now forces the code to be compiled
with .cpp files due to a circular dependency between AutoXChassis and
AutoXChassisBuilder