Implement PIDMotor Gyro + Timeout Fix + BackgroundMotorSystem fix
Modify PIDMotor such that the value of the inertial sensor's gyroscope
can be used instead of the encoder value; In practice it works just fine

AutoXChassis had a timeout bug due to std::chrono not working correctly;
The fix was to manually add the delay of the while loop to an
accumulator variable and use that to exit the loop.

Finally, there was a killer bug in BackgroundMotorSystem that caused
crashes: Any BackgroundMotor that detached from the system had the
"delete" keyword called on them even though they were already being
destroyed. This resulted in undefined behaviour, so delete the "delete"
keyword on that motor.