Increase accuracy of Chassis + Add Reset() to chassis builder
This commit can fully be built;

Increase accuracy of AutoChassis by adding the option to wait an amount
of time so it can settle within the margin of error of its target. Also,
never tare the rotation value of the gyro so that a steady state error
cannot make accuracy worse.

Add Reset() to chassis builder so that you can build more chassis.

This commit also fixes a dangling pointer in AutoXChassis.