Add Janky Camera System, ArmAssembly, and wrapper/
Modify inu::Motor to wrap pros::motor instead of using inheritance; The
reason for this is to provide better consistency between InuLib's
PascalCase interface and other wrapper classes.

wrapper/ is now a folder in inu/ that holds fundamental wrapper classes
that have a near 1:1 match between PROS' classes (albeit with extra
features). For instance, this commits introduces LineSensor and ADIMotor
as wrapper classes.

Added a meh camera system to track and rotate to aim towards a object,
though this is janky due to the object being tracked being sometimes
confused with the background.

Added ArmAssembly to have easier control over the arm and claw, and
added its respective builder class.