Add PID.h, two fields to PIDProfile, remove Reverse() in motors
Add PID.h to completely abstract away the PID algorithm from PIDMotor
and PIDInertialMotor to reduce code duplication. As each class had
different values to control the integral, add two new fields in
PIDProfile that directly set values for the max integral and the minimum
error needed to set it to 0.

Too scared to promise that reversing a motor will 100% work with the
current implemention, so completely remove it as it's not needed in this
competition.