Modify XLineFollower to use 5 sensors + FollowLine(distance)
Modify XLineFollower to use 5 line sensors instead of just 3 to reduce
the amount of setters/getters in the Builder and to decrease the
likelihood of missing the line completely when following it.

Add FollowLine(distance) to XLineFollower to follow a line over a
certain distance (encoder ticks from AutoXChassis) in accoradance with
our competition.