Massive AutoChassis API change + inu::port
Completely change the building process of AutoChassis and AutoXChassis
by passing in a subclass instead of filling in each variable
individually; The reasoning behind this is that every single option in
AutoChassis is optional and has a default that is not well expressed as
methods. In addition, AutoChassisBuilder simply had way too many simple
setters / getters. Passing in a class instead is much more reasonable
and gets the intention down.

Introduce inu::port so that no weird type errors occur; This will change
every instance of the library where a port is referenced, which
exception to inu::PIDVisionMotor as it is considered deprecated.

Some minor stuff as well:
* Add wrapper for Inertial Sensor
* Add Wifi mode to vision sensor
* Add exceptions to ArmAssemblyBuilder