Add Rebuild() functionality to all Chassis
Should you ever want to change the options of the chassis during
runtime, you can now do so with Rebuild(), which takes in an
AutoChassisBuilder instance and rebuilds the chassis. Mainly, we will be
using this to be able to change the direction of "forward" on the
chassis with a press of a button in order to use different arms on the
front and back of a robot.