Add Engine-centric Motor System
Add MechMotor, which is an interface to access the "Engine" that is
controlling a motor. "Engine" describes a class that takes control of a
motor. This flexibility allows it to swap out the following
functionality on the fly:

* Control by voltage
* Control by velocity
* Control by relative encoder positioning
* Control by absolute encoder positioning
* Control by Slew

The "SlewEngine" implemented fixed a critical bug in the original
SlewMotor, which was that Interpolate never returned negative numbers to
denote the opposite direction.