Delete all old code, move PID to util/
Delete all unused code, which includes the following:
* ArmAssembly
* ColorFollower
* XLineFollower
* AutoChassis
* AutoXChassis
* PIDMotor
* PIDIntertialMotor
* PIDVisionMotor
* SlewMotor

The reason for this is that the code for these is not the best as it
relies on instantiating and deleting motors to change from type to type,
which is fixed in MechMotor. However, these structures will not be used
in comp and will be implemented with a different approach, so they will
not be maintained either.