Add Field Controller + various bug fixes
Add Field Controller, which controls an XChassis with respect to the
field. It manipulates and controls the chassis to turn to or go to any
specific point on the field, given that the robot has a GPS sensor.
There are a ton of classes that were added:
* MathPoint / Field Point - Wrapper classes to transform field
  coordinates to classic math coordinates (angle, pos, and degree are
  different)
* Vector / VectorMath - Magnitude, with direction!

There are a ton of big fixes. One of the most major ones is not using
Slew to control the motors, as the heavy-load motors burned out without
them.